get keystore tests
