--- cmd/mailchain/internal/config/generate.go
@@ -1,3 +1,0 @@
-package config
-
-//go:generate mockgen -source=sent_store.go -package=configtest -destination=./configtest/sent_store_mock.go
