+++ cmd/mailchain/internal/http/server.go
@@ -57,7 +57,7 @@ func CreateRouter(cmd *cobra.Command) (http.Handler, error) {
 	if err != nil {
 		return nil, errors.WithMessage(err, "Could not config mailbox store")
 	}
-	keystore, err := config.GetKeystore()
+	keystore, err := config.DefaultKeystore().Get()
 	if err != nil {
 		return nil, errors.WithMessage(err, "could not create `keystore`")
 	}
